bilateralFilter implementation moved to separate file
